requirements.txt
astroid==2.8.5
atomicwrites==1.4.0
attrs==21.2.0
autopep8==1.6.0
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.7
colorama==0.4.4
Django==2.2.16
django-csvimport==2.16
django-filter==21.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
gunicorn==20.0.4
idna==3.3
iniconfig==1.1.1
isort==5.10.1
lazy-object-proxy==1.6.0
mccabe==0.6.1
mypy==0.910
mypy-extensions==0.4.3
packaging==21.2
platformdirs==2.4.0
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.10.0
pycodestyle==2.8.0
PyJWT==2.3.0
pylint==2.11.1
pyparsing==2.4.7
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dotenv==0.19.2
pytz==2021.3
requests==2.26.0
sqlparse==0.4.2
toml==0.10.2
typing-extensions==3.10.0.2
urllib3==1.26.7
wrapt==1.13.3
